feat: add API for category and categoryFilter

+++ app/src/main/java/com/group25/ecommercefashionapp/data/ProductColor.java
@@ -1,25 +1,25 @@
 package com.group25.ecommercefashionapp.data;
 
 public class ProductColor extends Item {
-    private final int product_id;
+    private final Long product_id;
     private final String hexColor;
-    private final int colorId;
+    private final Long colorId;
 
-    public ProductColor(int product_id, String hexColor, int colorId) {
+    public ProductColor(Long product_id, String hexColor, Long colorId) {
         super("");
         this.product_id = product_id;
         this.hexColor = hexColor;
         this.colorId = colorId;
     }
 
-    public ProductColor(int product_id, String hexColor, String name) {
+    public ProductColor(Long product_id, String hexColor, String name) {
         super(name);
-        this.colorId = -1;
+        this.colorId = -1L;
         this.hexColor = hexColor;
         this.product_id = product_id;
     }
 
-    public int getProduct_id() {
+    public Long getProduct_id() {
         return product_id;
     }
 
@@ -27,7 +27,7 @@ public class ProductColor extends Item {
         return hexColor;
     }
 
-    public int getColorId() {
+    public Long getColorId() {
         return colorId;
     }
